Rename package 'use_cases' to 'service'
