+++ dj_link/adapters/gateway.py
@@ -8,7 +8,7 @@ from typing import Iterable
 from dj_link.entities.custom_types import Identifier
 from dj_link.entities.link import Link, create_link
 from dj_link.entities.state import Commands, Components, Processes, Update
-from dj_link.use_cases.gateway import LinkGateway
+from dj_link.service.gateway import LinkGateway
 
 from .custom_types import PrimaryKey
 from .facade import DJAssignments, DJLinkFacade, DJProcess
